requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# Frontend Dependencies
streamlit==1.28.0
requests==2.31.0
pandas==2.1.3
plotly==5.18.0

# AI/ML Dependencies
tensorflow==2.14.0
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.26.2
scikit-learn==1.3.2 